Implemented call to pricing-service using Eureka for endpoint discovery
I had to update the spring boot from 2.1.5 to 2.4.1 in order to integrate the Eureka client. As a result, I also had to switch from Junit4 to Junit5, use new spring-hateoas library and new Swagger library. Due to the new Swagger library, the Swagger UI is now at http://localhost:8080/swagger-ui/ instead of the older URL.